Added device path functionality
In the future, to be able to modularise Turtle, the application will need to be split across multiple asm/efi files.
The only way of doing that is to learn how to call one file from another.
I am partway through this work, and this commit is more like a progress report.
Once it works, I probably won't need the device path utilities protocol, but we'll see when it happens :)

Also, a lot of the scripts have been improved!